+++ src/Ydb.Sdk/CHANGELOG.md
@@ -1,3 +1,5 @@
+## v0.28.0
+
 - Feat Reader/Writer: Added ConnectionString-based initialization.
 - Feat ADO.NET: Added support for `SnapshotRW` level isolation.
 - Feat: Added support for Yandex.Cloud metadata credentials (`EnableMetadataCredentials`) in `YdbConnectionStringBuilder` / connection string.